2478: Moved functionality into services
